divide up api v2 tests
